+++ code/back-end/src/main/java/met/cs673/team1/domain/entity/User.java
@@ -19,7 +19,7 @@ public class User {
 
     private String email;
 
-    private Integer budget;
+    private Double budget;
 
     @Column(name = "first_name")
     private String firstName;
